fix: exclude spec files from tsc build; move specs into utils/ for CI coverage
- packages/k8s/tsconfig.json: exclude src/**/*.spec.ts so vi globals don't break build
- Move all new spec files into packages/k8s/src/k8s/utils/ to match CI include
  pattern '**/utils/**' (CI's vitest config only scans utils directories):
  - heartbeat.spec.ts: tests parsePositiveMsEnv + WebSocketHeartbeat
  - k8s-index.spec.ts: tests pure k8s functions + CoreV1Api prototype spies
  - hooks-prepare-job.spec.ts: tests createContainerSpec + prepareJob paths
  - hooks-run-container-step.spec.ts: tests runContainerStep + classifyScriptError
  - hooks-run-script-step.spec.ts: tests formatScriptError + runScriptStep paths
