+++ packages/k8s/src/k8s/utils/heartbeat.spec.ts
@@ -1,4 +1,4 @@
-import { parsePositiveMsEnv, WebSocketHeartbeat } from './heartbeat'
+import { parsePositiveMsEnv, WebSocketHeartbeat } from '../heartbeat'
 
 vi.mock('@actions/core', () => ({
   debug: vi.fn(),
