+++ packages/k8s/tsconfig.json
@@ -5,8 +5,12 @@
       "outDir": "./lib",
       "rootDir": "./src"
     },
-    "esModuleInterop": true, /* Emit additional JavaScript to ease support for importing CommonJS modules. This enables 'allowSyntheticDefaultImports' for type compatibility. */
+    "esModuleInterop": true,
     "include": [
-      "src/**/*",
+      "src/**/*"
+    ],
+    "exclude": [
+      "src/**/*.spec.ts",
+      "node_modules"
     ]
 }
